+++ internal/ceres/dense_sparse_matrix.cc
@@ -48,9 +48,9 @@ DenseSparseMatrix::DenseSparseMatrix(const TripletSparseMatrix& m)
   const double* values = m.values();
   const int* rows = m.rows();
   const int* cols = m.cols();
-  int num_nonzeros = m.num_nonzeros();
+  int64_t num_nonzeros = m.num_nonzeros();
 
-  for (int i = 0; i < num_nonzeros; ++i) {
+  for (int64_t i = 0; i < num_nonzeros; ++i) {
     m_(rows[i], cols[i]) += values[i];
   }
 }
@@ -98,7 +98,9 @@ int DenseSparseMatrix::num_rows() const { return m_.rows(); }
 
 int DenseSparseMatrix::num_cols() const { return m_.cols(); }
 
-int DenseSparseMatrix::num_nonzeros() const { return m_.rows() * m_.cols(); }
+int64_t DenseSparseMatrix::num_nonzeros() const {
+  return m_.rows() * m_.cols();
+}
 
 const Matrix& DenseSparseMatrix::matrix() const { return m_; }
