+++ internal/ceres/triplet_sparse_matrix.h
@@ -51,7 +51,7 @@ namespace ceres::internal {
 class CERES_NO_EXPORT TripletSparseMatrix final : public SparseMatrix {
  public:
   TripletSparseMatrix();
-  TripletSparseMatrix(int num_rows, int num_cols, int max_num_nonzeros);
+  TripletSparseMatrix(int num_rows, int num_cols, int64_t max_num_nonzeros);
   TripletSparseMatrix(int num_rows,
                       int num_cols,
                       const std::vector<int>& rows,
@@ -76,18 +76,18 @@ class CERES_NO_EXPORT TripletSparseMatrix final : public SparseMatrix {
   // clang-format off
   int num_rows()        const final   { return num_rows_;     }
   int num_cols()        const final   { return num_cols_;     }
-  int num_nonzeros()    const final   { return num_nonzeros_; }
+  int64_t num_nonzeros()    const final   { return num_nonzeros_; }
   const double* values()  const final { return values_.get(); }
   double* mutable_values() final      { return values_.get(); }
   // clang-format on
-  void set_num_nonzeros(int num_nonzeros);
+  void set_num_nonzeros(int64_t num_nonzeros);
 
   // Increase max_num_nonzeros and correspondingly increase the size
   // of rows_, cols_ and values_. If new_max_num_nonzeros is smaller
   // than max_num_nonzeros_, then num_non_zeros should be less than or
   // equal to new_max_num_nonzeros, otherwise data loss is possible
   // and the method crashes.
-  void Reserve(int new_max_num_nonzeros);
+  void Reserve(int64_t new_max_num_nonzeros);
 
   // Append the matrix B at the bottom of this matrix. B should have
   // the same number of columns as num_cols_.
@@ -102,7 +102,7 @@ class CERES_NO_EXPORT TripletSparseMatrix final : public SparseMatrix {
   void Resize(int new_num_rows, int new_num_cols);
 
   // clang-format off
-  int max_num_nonzeros() const { return max_num_nonzeros_; }
+  int64_t max_num_nonzeros() const { return max_num_nonzeros_; }
   const int* rows()      const { return rows_.get();       }
   const int* cols()      const { return cols_.get();       }
   int* mutable_rows()          { return rows_.get();       }
@@ -148,8 +148,8 @@ class CERES_NO_EXPORT TripletSparseMatrix final : public SparseMatrix {
 
   int num_rows_;
   int num_cols_;
-  int max_num_nonzeros_;
-  int num_nonzeros_;
+  int64_t max_num_nonzeros_;
+  int64_t num_nonzeros_;
 
   // The data is stored as three arrays. For each i, values_[i] is
   // stored at the location (rows_[i], cols_[i]). If the there are
